Implement TXT records for use in skydns
